+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- [WIP] The ability to skip data inconsistency errors when reading with the `IndexedDBAdapter` adapter.
+
 ## [2.5.0-beta.2] - 2020-06-22
 ### Added
 - Store